+++ infrastructure/__init__.py
@@ -2,5 +2,11 @@
 from .container import DependencyContainer
 from .eventbus import InMemoryEventBus
 from .graph_builder import InMemoryGraphBuilder
+from .config_loader import ConfigLoader
 
-__all__ = ["DependencyContainer", "InMemoryEventBus", "InMemoryGraphBuilder"]
+__all__ = [
+    "DependencyContainer",
+    "InMemoryEventBus",
+    "InMemoryGraphBuilder",
+    "ConfigLoader",
+]